requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
hetzner-dns-tools==0.0.12
idna==3.4
miniupnpc==2.0.2
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
